% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{default_parse_fun}
\alias{default_parse_fun}
\title{Default Parse Function}
\usage{
default_parse_fun(verbose)
}
\arguments{
\item{verbose}{If \code{TRUE}, display descriptive message.}
}
\value{
Reference to function \code{\link[readr:read_csv]{readr::read_csv()}}
}
\description{
Return the default file parsing function, and throw error
if it is not available.
}
\keyword{internal}
