% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{download_file}
\alias{download_file}
\title{Download File}
\usage{
download_file(url, fname = basename(url), download_dir = tempdir())
}
\arguments{
\item{url}{The url of the file to download.}

\item{fname}{The name of the file to save to. By default, the
filename at the end of the url is used.}

\item{download_dir}{The directory to place the downloaded file in.
By default, the session temporary directory is used.}
}
\value{
The path of the downloaded file.
}
\description{
Helper function to download files using curl::curl_download().
This function can be supplied in place of argument \code{parse_fun}
to download files without attempting to read their contents.
}
\examples{
\dontrun{
download_file(paste0("https://raw.githubusercontent.com/",
  "InteragencyEcologicalProgram/Status-and-Trends/master/data/",
  "Grandtab_adultsalmon.csv"))
}

}
