% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{get_baydeltalive_data}
\alias{get_baydeltalive_data}
\title{Download Bay Delta Live Data}
\usage{
get_baydeltalive_data(
  asset_id,
  path_suffix,
  fnames,
  parse_fun,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{asset_id}{The asset identifier.}

\item{path_suffix}{Path suffix(es), specifying subfolders of the
asset to search}

\item{fnames}{A vector of file names in the package to download.
Supports regex.}

\item{parse_fun}{A function to parse datasets. Default assumes that
all files in fnames can be parsed using \code{readr::read_csv()}.}

\item{...}{Additional arguments to pass to \code{parse_fun}.}

\item{verbose}{If \code{TRUE}, display descriptive messages.}
}
\value{
a named list of dataframes. The list also includes an
attribute "Notes" of same length containing the notes section
extracted from the report files.
}
\description{
Download data from \href{https://cbr.washington.edu}{Bay Delta Live}.
}
\examples{
\dontrun{
 parse_remote_excel = function(url, type = c(".xlsx", ".xls"), ...) {
   type = match.arg(type, c(".xlsx", ".xls"))
   tf = tempfile(fileext = type)
   curl::curl_download(url, tf)
   readxl::read_excel(tf, ...)
}
get_baydeltalive_data("00c870b6fdc0e30d0f92d719984cfb44",
  "application/vnd.ms-excel", "Field_Data_[0-9]{4}-[0-9]{4}x*",
  parse_remote_excel, guess_max = 100000L)
}

}
