% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{get_edi_data}
\alias{get_edi_data}
\title{Download EDI Package Files}
\usage{
get_edi_data(pkg_id, fnames, parse_fun, ..., verbose = TRUE)
}
\arguments{
\item{pkg_id}{The EDI Package ID}

\item{fnames}{A vector of file names in the package to download.
Supports regex.}

\item{parse_fun}{Function to parse datasets. Default assumes that
all files in fnames can be parsed using \code{readr::read_csv()}.}

\item{...}{Additional arguments to pass to \code{parse_fun}.}

\item{verbose}{If TRUE, display descriptive messages.}
}
\value{
a list of dataframes
}
\description{
Download the latest revision of files in an EDI package
}
