% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{get_ftp_data}
\alias{get_ftp_data}
\title{Download FTP Data}
\usage{
get_ftp_data(ftp_address, dir_path, fnames, parse_fun, ..., verbose = TRUE)
}
\arguments{
\item{ftp_address}{The FTP server address.}

\item{dir_path}{FTP directory to search.}

\item{fnames}{A vector of file names in the package to download.
Supports regex.}

\item{parse_fun}{A function to parse datasets. Default assumes that
all files in fnames can be parsed using \code{readr::read_csv()}.}

\item{...}{Additional arguments to pass to \code{parse_fun}.}

\item{verbose}{If \code{TRUE}, display descriptive messages.}
}
\value{
a named list of dataframes. The list also includes an
attribute "Notes" of same length containing the notes section
extracted from the report files.
}
\description{
Download data from an FTP server.
}
\examples{
\dontrun{
get_baydeltalive_data("00c870b6fdc0e30d0f92d719984cfb44",
  "application/vnd.ms-excel", "Field_Data_[0-9]{4}-[0-9]{4}x*",
  parse_remote_excel, guess_max = 100000L)
}

}
