% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{get_odp_data}
\alias{get_odp_data}
\title{Download Open Data Portal Package Files}
\usage{
get_odp_data(
  portal_url = "https://data.cnra.ca.gov",
  pkg_id,
  fnames,
  parse_fun,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{portal_url}{The base URL of the portal, e.g. \code{"data.cnra.ca.gov"}.}

\item{pkg_id}{The package identifier.}

\item{fnames}{A vector of file names in the package to download.
Supports regex.}

\item{parse_fun}{A function to parse datasets. Default assumes that
all files in fnames can be parsed using \code{readr::read_csv()}.}

\item{...}{Additional arguments to pass to \code{parse_fun}.}

\item{verbose}{If \code{TRUE}, display descriptive messages.}
}
\value{
a named list of dataframes.
}
\description{
Download files from an Open Data Portal, assuming it uses the CKAN
API.
}
