% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{get_redbluff_data}
\alias{get_redbluff_data}
\title{Download Redbluff Data}
\usage{
get_redbluff_data(
  report_year,
  start_year = 2004,
  parse_fun,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{report_year}{The report year.}

\item{start_year}{The initial year to retrieve data for.
Default is 2004.}

\item{parse_fun}{A function to parse datasets. Default assumes that
all files in fnames can be parsed using `readr::read_csv()`.}

\item{...}{Additional arguments to pass to `parse_fun`.}

\item{verbose}{If `TRUE`, display descriptive messages.}
}
\value{
a named list of dataframes. The list also includes an
  attribute "Notes" of same length containing the notes section
  extracted from the report files.
}
\description{
Download Redbluff data from [cbr.washington.edu](https://cbr.washington.edu).
}
\examples{
\dontrun{
get_redbluff_data(2018, 2016, na = "--")
}

}
