% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{lt_avg_line}
\alias{lt_avg_line}
\title{Add a horizontal line to a ggplot}
\usage{
lt_avg_line(lt_avg)
}
\arguments{
\item{lt_avg}{The long-term average of the data}
}
\value{
A horizontal dashed red line to a ggplot
}
\description{
Adds a horizontal dashed red line to a ggplot. Its intended use
    is to represent the provided long-term average of a dataset. For now the
    user is required to input the long-term average value into the function.
}
