% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{parse_ftp_index}
\alias{parse_ftp_index}
\title{Parse FTP Index}
\usage{
parse_ftp_index(url)
}
\arguments{
\item{url}{The url of the FTP directory to parse.}
}
\value{
A character vector of file names.
}
\description{
Parse an FTP index page.
}
\keyword{internal}
