% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{parse_html_index}
\alias{parse_html_index}
\title{Parse HTML Index}
\usage{
parse_html_index(url, node = "//a", attribute = "href")
}
\arguments{
\item{url}{The url of the page to parse.}

\item{node}{The HTML node to extract. Default is link anchor,
i.e. \verb{<a>}.}

\item{attribute}{if not \code{NULL}, the attribute of the node to
extract.}
}
\value{
A character vector of nodes or attributes.
}
\description{
Parse an HTML index page.
}
\examples{
\dontrun{
parse_html_index(paste0("https://emp.baydeltalive.com/assets/",
  "00c870b6fdc0e30d0f92d719984cfb44/application/vnd.ms-excel"))
}

}
\keyword{internal}
