% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{round_to_mult}
\alias{round_to_mult}
\title{Round an integer to the nearest multiple}
\usage{
round_to_mult(num, mult)
}
\arguments{
\item{num}{Integer to round}

\item{mult}{Mulitple value to round to}
}
\value{
A value rounded to the nearest multiple of the \code{mult} parameter
}
\description{
Used to round a year value to the nearest multiple of a defined value. This is
    an internal function used to define x-axis breaks for plots.
}
\keyword{internal}
