% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_join.R
\name{schema_join}
\alias{schema_join}
\title{Applies the joins in accordance to a relationship schema.}
\usage{
schema_join(schema, data, start = NULL, returnSchema = F)
}
\arguments{
\item{schema}{A schema table that mirrors the structure of an Access schema.}

\item{data}{A list of data tables to be joined}

\item{start}{The starting relational table to begin construction. If NULL, will start at the first table specified by the schema.}

\item{returnSchema}{Return the schema table as modified by the function prior to the joining process. Internal use only.}
}
\value{

}
\description{
Applies the joins in accordance to a relationship schema.
}
