% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smonitr-package.R
\docType{package}
\name{smonitr}
\alias{smonitr}
\alias{smonitr-package}
\title{IEP Seasonal Monitoring Report Tools}
\description{
A collection of R functions for the IEP Seasonal Monitoring Reports.
}
\author{
\strong{Maintainer}: Dave Bosworth \email{David.Bosworth@water.ca.gov} (\href{https://orcid.org/0000-0003-0740-3390}{ORCID})

}
\keyword{internal}
