% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{smr_theme}
\alias{smr_theme}
\title{Custom formatting for IEP Seasonal Monitoring Report figures}
\usage{
smr_theme()
}
\value{
A ggplot theme.
}
\description{
This custom theme builds off of the \verb{[ggplot2::theme_bw()]}
theme to remove plot gridlines and modify the text size of
various plot elements.
}
