% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_templates.R
\name{standard_column_plot}
\alias{standard_column_plot}
\title{Column plot template for IEP Seasonal Monitoring Report figures}
\usage{
standard_column_plot(
  data,
  mapping,
  report_year,
  season = c("annual", "winter", "spring", "summer", "fall"),
  type = c("recent", "all"),
  ...
)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{report_year}{The user-defined report year for the Seasonal
Monitoring Report. Must be an integer.}

\item{season}{The season for the Seasonal Monitoring Report.
Must be one of the following: \code{"winter"}, \code{"spring"},
\code{"summer"}, \code{"fall"}, or \code{"annual"}.}

\item{type}{The scale type to use. \code{type = "all"} defines the
minimum x-axis limit as 1966, which corresponds to the earliest
survey shown in the fall season report (Fall Midwater Trawl).
\code{type = "recent"} defines the minimum x-axis limit as 2004, which
corresponds to the Pelagic Organism Decline.}

\item{...}{other arguments to pass to \verb{[geom_col()]}}
}
\description{
A template function for producing column plots for the IEP
Seasonal Monitoring Report.
}
