% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geoms.R
\docType{data}
\name{StatLTavg}
\alias{StatLTavg}
\title{StatLTavg ggproto}
\format{An object of class \code{StatLTavg} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatLTavg
}
\description{
StatLTavg ggproto
}
\keyword{internal}
