% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geoms.R
\docType{data}
\name{StatMissing}
\alias{StatMissing}
\title{StatMissing ggproto}
\format{
An object of class \code{StatMissing} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatMissing
}
\description{
StatMissing ggproto
}
\keyword{internal}
