% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{std_x_axis_label}
\alias{std_x_axis_label}
\title{Add a standardized x-axis label to a ggplot}
\usage{
std_x_axis_label(season = c("winter", "spring", "summer", "fall", "annual"))
}
\arguments{
\item{season}{The user defined season for the Seasonal Monitoring Report. Must be
one of the following: \code{"winter", "spring", "summer", "fall", "annual"}.}
}
\value{
A standardized x-axis label to a ggplot object. Each season option
    provides a label in the format "Year (Month - Month)". The argument
    \code{season = "annual"} provides the label "Year" without specified months.
}
\description{
Adds a standardized x-axis label to a ggplot based on the season
    of the report.
}
