% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{std_x_axis_rec_years}
\alias{std_x_axis_rec_years}
\title{Standardize x-axis elements for ggplot of recent years}
\usage{
std_x_axis_rec_years(rpt_yr)
}
\arguments{
\item{rpt_yr}{The user-defined report year for the Seasonal Monitoring Report.
Must be an integer.}
}
\value{
A ggplot layer that defines the x-axis limits and breaks for plots of
    recent data. The default limits are \code{rpt_yr - 14} to \code{rpt_yr},
    and the default break interval is 5 years.
}
\description{
Standardizes the x-axis limits and breaks for a ggplot of recent
    years of data. "Recent years" is defined as the prior 15 years from the
    \code{rpt_yr}. For plots with \code{geom_col} geoms, the x-variable used
    in the ggplot needs to be a \strong{factor}.
}
