% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_access.R
\name{bridge_access}
\alias{bridge_access}
\title{Create the connection to an Access database and pull the requested tables.}
\usage{
bridge_access(file, tables = "check", out = tempdir(), ...)
}
\arguments{
\item{file}{File path to the Access database file.}

\item{tables}{A vector of table names to pull. This can be left blank to provide a list of options}

\item{out}{As needed. File path to save the outputted rds file if a 32-bit R is launched from the terminal.}
}
\value{

}
\description{
Create the connection to an Access database and pull the requested tables.
}
\examples{
\dontrun{

}

}
