% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{choose_files}
\alias{choose_files}
\title{Choose Files}
\usage{
choose_files(files, selections = ".*", verbose)
}
\arguments{
\item{files}{A vector of file names.}

\item{selections}{A vector of file names to choose from \code{files}.
Supports regex.}

\item{verbose}{If \code{TRUE}, display descriptive message.}
}
\value{
A vector of file names.
}
\description{
Subset a list of files based on filenames or regex strings.
}
