% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{extract_files}
\alias{extract_files}
\title{Extract Files From Zip}
\usage{
extract_files(path, fnames = ".*", ..., verbose = TRUE)
}
\arguments{
\item{path}{The path to the zipfile.}

\item{fnames}{A vector of file names in the archive to extract.
Supports regex.}

\item{...}{Other arguments passed to \code{utils::unzip()}.}

\item{verbose}{If \code{TRUE}, display descriptive message.}
}
\value{
A vector of extracted file paths.
}
\description{
Extract files from a zip archive.
}
\examples{
\dontrun{
f = download_file("ftp://ftp.wildlife.ca.gov/Delta\%20Smelt/SKT.zip")
extract_files(f, "SKT", exdir = tempdir())
}

}
