% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{get_edi_data}
\alias{get_edi_data}
\title{Download EDI Package Files}
\usage{
get_edi_data(pkg_id, fnames, parse_fun, ..., verbose = TRUE)
}
\arguments{
\item{pkg_id}{The package identifier.}

\item{fnames}{A vector of file names in the package to download.
Supports regex.}

\item{parse_fun}{A function to parse datasets. Default assumes that
all files in fnames can be parsed using \code{readr::read_csv()}.}

\item{...}{Additional arguments to pass to \code{parse_fun}.}

\item{verbose}{If \code{TRUE}, display descriptive messages.}
}
\value{
A named list of dataframes.
}
\description{
Download the latest revision of files in an EDI package
}
\examples{
\dontrun{
get_edi_data(233, "YBFMP_Fish", guess_max = 1000000)
}

}
