% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{get_grandtab_data}
\alias{get_grandtab_data}
\title{Download GrandTab Data}
\usage{
get_grandtab_data(
  season = c("Winter", "Spring", "Fall", "Late-Fall"),
  parse_fun,
  ...,
  verbose = TRUE
)
}
\description{
Download CDFW GrandTab data from SacPass.
}
