% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{get_redbluff_data}
\alias{get_redbluff_data}
\title{Download Redbluff Data}
\usage{
get_redbluff_data(
  report_year,
  start_year = 2004,
  parse_fun,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{report_year}{The report year.}

\item{start_year}{The initial year to retrieve data for.
Default is \code{2004}.}

\item{parse_fun}{A function to parse datasets. Default assumes that
all files in fnames can be parsed using \code{\link[readr:read_csv]{readr::read_csv()}}.}

\item{...}{Additional arguments to pass to \code{parse_fun}.}

\item{verbose}{If \code{TRUE}, display descriptive messages.}
}
\value{
a list of dataframes, each element corresponds to a year
the sequence \code{start_year:report_year}. The list also
includes an attribute \code{"Notes"} of same length and order
containing the notes section extracted each report file.
}
\description{
Download Redbluff data from
\href{https://cbr.washington.edu}{cbr.washington.edu}.
}
\examples{
\dontrun{
get_redbluff_data(2018, 2016, na = "--")
}

}
