% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{missing_data_symb}
\alias{missing_data_symb}
\title{Add symbols to a ggplot representing missing data}
\usage{
missing_data_symb(df, yr_var, rpt_yr, symb_size)
}
\arguments{
\item{df}{The dataframe that contains the data in the ggplot}

\item{yr_var}{The name of the variable in \code{df} that contains the years
used in the ggplot. Should be a \strong{factor}.}

\item{rpt_yr}{The user-defined report year for the Seasonal Monitoring Report.
Must be an integer.}

\item{symb_size}{Specifies the size of the symbol used to represent missing
data. Must be numeric. Usually either \code{symb_size = 1} or
\code{symb_size = 2}.}
}
\value{
A ggplot object with tan triangle symbols that represent years with
    missing data.
}
\description{
Adds tan triangle symbols to a ggplot to represent years with
    missing data. This function is intended to be used on plots with
    \code{geom_col} geoms.
}
