% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{parse_remote_excel}
\alias{parse_remote_excel}
\title{Parse Remotely-Hosted Excel File}
\usage{
parse_remote_excel(path, ...)
}
\arguments{
\item{path}{The URL or FTP directory of the Excel file.}

\item{...}{Other arguments to pass to \code{\link[readxl:read_excel]{readxl::read_excel()}}}
}
\description{
Helper function for parsing an Excel file hosted on a website
or FTP server.
}
\note{
This function will be defunct once \code{readxl::read_excel()}
\href{https://github.com/tidyverse/readxl/issues/278}{supports reading from more general inputs}.
}
