% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{smr_y_axis}
\alias{smr_y_axis}
\title{Custom y-axis for IEP Seasonal Monitoring Report figures}
\usage{
smr_y_axis(expand = expand_scale(mult = c(0, 0.05)), ...)
}
\arguments{
\item{expand}{Vector of range expansion constants used to add some
padding around the data, to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[ggplot2:expand_scale]{expand_scale()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:scale_y_continuous]{ggplot2::scale_y_continuous()}}}
}
\value{
A ggplot continuous x-axis scale.
}
\description{
Standardizes the y-axis a ggplot. This is really just a wrapper
around\code{\link[ggplot2:scale_y_continuous]{ggplot2::scale_y_continuous()}} with a default setting
for the \code{expand} argument to reduce whitespace.
}
