% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{std_x_axis_all_years}
\alias{std_x_axis_all_years}
\title{Standardize x-axis elements for ggplot of all years}
\usage{
std_x_axis_all_years(
  rpt_yr,
  x_scale_type = c("discrete", "cont"),
  start_yr = 1966,
  break_int = 10
)
}
\arguments{
\item{rpt_yr}{The user-defined report year for the Seasonal Monitoring Report.
Must be an integer.}

\item{x_scale_type}{The scale type for the x variable. Can be either
\code{x_scale_type = "discrete"} or \code{x_scale_type = "cont"}. Plots
with \code{geom_col} geoms are "discrete" scale type, and \code{geom_line}
\code{geom_area} geoms are "cont" or continuous scale type.}

\item{start_yr}{The start year that defines the minimum x-axis limit for the
ggplot. Must be an integer. Default is 1966.}

\item{break_int}{The x-axis break interval in years. Must be an integer.
Default is 10.}
}
\value{
A ggplot layer that defines the x-axis limits and breaks based upon
    user-defined arguments.
}
\description{
Standardizes the x-axis limits and breaks for a ggplot of
    all years of data. For plots with a discrete x-variable including
    \code{geom_col} geoms, the x-variable used in the ggplot needs to
    be a \strong{factor}.
}
