% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{std_x_axis_rec_years}
\alias{std_x_axis_rec_years}
\title{Standardize x-axis elements for ggplot of recent years}
\usage{
std_x_axis_rec_years(rpt_yr)
}
\arguments{
\item{rpt_yr}{The user-defined report year for the Seasonal Monitoring Report.
Must be an integer.}
}
\value{
Two ggplot layers that define the x-axis limits and breaks based upon
user-defined arguments
}
\description{
Standardizes the x-axis limits and breaks for a ggplot of recent
years of data. "Recent years" is defined as the prior 15 years from the
\code{rpt_yr}.
}
\section{Special Instructions}{

The x-variable in the ggplot needs to be "numeric" type. This function
works best if the dataset is filtered to the necessary date range
beforehand.
}

